разнес ответственность моделей по разным фабрикам + создал в отдельном файле фабрику вопросов с генерацией случайного вопроса по индексу из массива
